requirements.txt
# Framework web
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0

# ORM et base de données
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
alembic==1.12.1

# Formulaires et validation
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Authentification et sécurité
Flask-Login==0.6.3
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0

# Traitement d'images et données
Pillow>=10.1.0
pandas==2.1.3
numpy==1.26.2

# Utilitaires
requests==2.31.0
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Serveur de production
gunicorn==21.2.0

# Tests
pytest==7.4.3
pytest-flask==1.3.0